feat: make prod running
