+++ server/index.ts
@@ -6,50 +6,84 @@ import cors from "cors";
 import dotenv from "dotenv";
 import http from "http";
 import { MongoClient, ServerApiVersion } from "mongodb";
-
+import PokemonMongo from "./src/datasource/mongo-api/index"; // Assuming this is your Mongo data source
+import PokemonAPI from "./src/datasource/pokemon-api/index"; // Assuming this is your external API data source
+import resolvers from "./src/resolvers"; // Your GraphQL resolvers
+import typeDefs from "./src/typeDefs"; // Your GraphQL type definitions
 
 dotenv.config();
+
 const port = process.env.PORT || 5844;
 const app = express();
+
+// MongoDB Connection URI
 const password = encodeURIComponent(process.env.SECRET_MONGODB);
 const uri = `mongodb+srv://pokemonsv2:${password}@pokemonsv2.uczp40b.mongodb.net/?retryWrites=true&w=majority&appName=Pokemonsv2`;
 
+const client = new MongoClient(uri, {
+    serverApi: {
+        version: ServerApiVersion.v1,
+        strict: true,
+        deprecationErrors: true,
+    },
+});
 
+const httpServer = http.createServer(app);
 
-//middlewares
+// Middleware
 app.use(express.json());
 app.use(cors());
 
-//mongo URI
-const client = new MongoClient(uri, {
-    serverApi: {
-      version: ServerApiVersion.v1,
-      strict: true,
-      deprecationErrors: true,
-    },
-  });
-
-  const run = async () => {
-    try{
+const run = async () => {
+    try {
         await client.connect();
-
-         await client.db("admin").command({ ping: 1 });
-        console.log(
-          "Pinged your deployment. You successfully connected to MongoDB!"
-        );
-    }
-    finally{
-
-    }
+        console.log("Connected to MongoDB successfully");
+    } catch (err) {
+        console.error("Failed to connect to MongoDB", err);
+        process.exit(1); // Exit process if the connection fails
     }
 
-    run().catch(error => console.log)
+    const server = new ApolloServer({
+        typeDefs,
+        resolvers,
+        plugins: [ApolloServerPluginDrainHttpServer({ httpServer })],
+    });
 
-    app.get('/',(req,res)=>{
-        res.send('Car Junction Backend Server Running...')
-    })
+    await server.start();
 
-    app.listen(port,()=>{
-        console.log(console.log(`Server is running on port ${port}`))
-    })
+    app.use(
+        "/graphql",
+        cors(), // Apply CORS middleware
+        express.json(), // Apply JSON middleware
+        expressMiddleware(server, {
+            context: async ({ req }) => {
+                const { cache } = server;
 
+                return {
+                    token: req.headers.token,
+                    dataSources: {
+                        pokemonAPI: new PokemonAPI({ cache }), // External API Data Source
+                        basePokemons: new PokemonMongo({
+                            modelOrCollection: client.db("pokemons").collection("ids"), // MongoDB Collection
+                        }),
+                    },
+                };
+            },
+        })
+    );
+
+    // REST Endpoint
+    app.get('/', (req, res) => {
+        res.send('Car Junction Backend Server Running...');
+    });
+
+    // Start the HTTP server
+    await new Promise((resolve) =>
+        httpServer.listen({ port }, resolve)
+    );
+
+    console.log(`🚀 Server ready at http://localhost:${port}/graphql`);
+};
+
+// Run the server
+run().catch(error => console.error(error));
